# spec/openapi.yaml
openapi: 3.0.4
info:
  title: Reiseapp API
  version: 0.1.0
  description: API-kontrakt for reiseappen.
tags:
  - name: Health
    description: Livssjekk for applikasjonen.
  - name: Person
    description: Oppretting og uthenting av personer.
  - name: Room
    description: Uthenting av rom og romdetaljer.
paths:
  /health:
    get:
      tags: [Health]
      operationId: health
      summary: Kontrollerer at API-et kan svare.
      description: Livssjekk for applikasjonen. Kontrollerer ikke databasetilkoblingen.
      responses:
        '200':
          description: API-et svarer.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
  /persons:
    get:
      tags: [Person]
      operationId: listPersons
      summary: Henter alle registrerte personer.
      responses:
        '200':
          description: Alle registrerte personer, sortert på ID.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PersonResponse'
    post:
      tags: [Person]
      operationId: createPerson
      summary: Registrerer en ny person.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePersonRequest'
      responses:
        '201':
          description: Personen ble registrert.
          headers:
            Location:
              description: Relativ URI til den registrerte personen.
              schema:
                type: string
                format: uri-reference
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonResponse'
        '400':
          description: Forespørselen er ugyldig.
  /persons/{personId}:
    get:
      tags: [Person]
      operationId: getPerson
      summary: Henter én registrert person.
      parameters:
        - name: personId
          in: path
          required: true
          description: Personens ID.
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '200':
          description: Personen som har angitt ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonDetailsResponse'
        '404':
          description: Personen finnes ikke.
  /rooms:
    get:
      tags: [Room]
      operationId: listRooms
      summary: Henter alle rom.
      responses:
        '200':
          description: Alle rom, sortert på romnummer.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoomResponse'
  /rooms/{roomNumber}:
    get:
      tags: [Room]
      operationId: getRoom
      summary: Henter detaljer om ett rom.
      parameters:
        - name: roomNumber
          in: path
          required: true
          description: Romnummeret.
          schema:
            type: integer
            format: int32
            minimum: 1
      responses:
        '200':
          description: Rommet med angitt romnummer.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomDetailsResponse'
        '400':
          description: Romnummeret er ugyldig.
        '404':
          description: Rommet finnes ikke.
components:
  schemas:
    PersonDetailsResponse:
      allOf:
        - $ref: '#/components/schemas/PersonResponse'
        - type: object
          properties:
            assignedRoom:
              description: Tildelt rom. Null eller utelatt hvis personen ikke har rom.
              nullable: true
              allOf:
                - $ref: '#/components/schemas/RoomResponse'
    RoomDetailsResponse:
      allOf:
        - $ref: '#/components/schemas/RoomResponse'
        - type: object
          required: [persons]
          properties:
            persons:
              description: Tildelte personer sortert på ID. Tom liste for ledige rom.
              type: array
              items:
                $ref: '#/components/schemas/PersonResponse'
    HealthResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [UP]
          example: UP
    CreatePersonRequest:
      type: object
      required: [name, department, email, phoneNumber, gender]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
          example: Vennlige Foss
        department:
          type: string
          minLength: 1
          maxLength: 100
          example: Produkt
        email:
          type: string
          format: email
          maxLength: 254
          example: vennlige.foss@reiseapp.test
        phoneNumber:
          type: string
          minLength: 1
          maxLength: 32
          example: +47 0000 0011
        gender:
          type: string
          minLength: 1
          maxLength: 50
          example: kvinne
    PersonResponse:
      type: object
      required: [id, name, department, email, phoneNumber, gender, registrationDate]
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
          example: 11
        name:
          type: string
          minLength: 1
          maxLength: 200
          example: Vennlige Foss
        department:
          type: string
          minLength: 1
          maxLength: 100
          example: Produkt
        email:
          type: string
          format: email
          maxLength: 254
          example: vennlige.foss@reiseapp.test
        phoneNumber:
          type: string
          minLength: 1
          maxLength: 32
          example: +47 0000 0011
        gender:
          type: string
          minLength: 1
          maxLength: 50
          example: kvinne
        registrationDate:
          type: string
          format: date
          example: '2026-09-15'
    RoomResponse:
      type: object
      required: [roomNumber, sizeSquareMeters, numberOfBeds, hasBalcony, lastRenovatedYear]
      properties:
        roomNumber:
          type: integer
          format: int32
          minimum: 1
          example: 204
        sizeSquareMeters:
          type: integer
          format: int32
          minimum: 1
          example: 28
        numberOfBeds:
          type: integer
          format: int32
          enum: [1, 2, 4]
          example: 2
        hasBalcony:
          type: boolean
          example: true
        lastRenovatedYear:
          type: integer
          format: int32
          minimum: 1900
          maximum: 2100
          example: 2024